requirements.txt
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
boltons==21.0.0
bracex==2.5.post1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-option-group==0.5.6
colorama==0.4.6
cssselect2==0.7.0
defusedxml==0.7.1
Deprecated==1.2.14
exceptiongroup==1.2.2
face==22.0.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.54.1
glom==22.1.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
html5lib==1.1
idna==3.10
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
packaging==24.1
pdfkit==1.0.0
peewee==3.17.7
pillow==11.0.0
protobuf==4.25.5
pycparser==2.22
pydyf==0.11.0
Pygments==2.18.0
pyphen==0.16.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.5.3
rpds-py==0.20.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.12
semgrep==1.92.0
setuptools==75.2.0
six==1.16.0
SQLAlchemy==2.0.36
tinycss2==1.3.0
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.3
wcmatch==8.5.2
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.16.0
WTForms==3.2.0
yara-python==4.5.1
zipp==3.20.2
zopfli==0.2.3.post1
